requirements.txt
aerich==0.5.3
aiofiles==0.7.0
aiomysql==0.0.21
aiosqlite==0.16.1
attrs==21.2.0
cattrs==1.7.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
ddlparse==1.9.0
dictdiffer==0.8.1
httptools==0.2.0
idna==2.10
importlib-metadata==4.5.0
iso8601==0.1.14
jsonschema==3.2.0
jsonschema-extractor==0.9
lark-parser==0.11.3
multidict==5.1.0
numpy==1.20.3
opencv-python==4.5.2.54
Pillow==8.2.0
pydantic==1.8.2
PyMySQL==0.9.3
pyparsing==2.4.7
pypika-tortoise==0.1.1
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
sanic==21.3.4
sanic-openapi==21.3.2
sanic-routing==0.6.2
schematics==2.1.0
selenium==3.141.0
six==1.16.0
swagger-schema==0.5.1
tortoise-orm==0.17.4
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.5
uvloop==0.15.2
websockets==8.1
zipp==3.4.1
